test: add bundlesize
This will let us prevent regressions in bundlesize.

I chose to monitor the bundlesize of webpack and rollup build because
this is what ultimately matters to the user. If we break something in
our library that would e.g. break tree shaking in one of these bundlers
we would want to notice.